moved comments into a separate module
