+++ js/const.js
@@ -1,0 +1,4 @@
+const INIT_COMMENTS_NUMBER = 5;
+const COUNT_STEP = 5;
+
+export {INIT_COMMENTS_NUMBER, COUNT_STEP};
